Deprecate Demux
Originally, FFNet always overwrote gradient when it backpropagated the
error. But this made me angry because of Demux.

If one FFVariable is connected more than one modules, it must add
gradients come from connected modules. But overwritting modules doesn't
know whether connected FFVariable is independant to other modules or
not. Demux solved this problem.

But using Demux is really complicated and annoying. I have to set every
demuxing variables's count and their pointers. So I decided to change
them by adding gradient, not overwritting.

Because of this most of the code in the FFModule and FFNet have been
changed.